Face-recognization-attandance-system
This project introduces a Face Detection Attendance
System designed to automate the process of attendance tracking.
Arduino-based hardware setup Python for face detection and
recognition enhance the accuracy and efficiency of the system.
The implemented solution begins with capturing faces using a
webcam, and these images are then processed using Python’s
Open-CV library. The face detection algorithm identifies and
recognizes individuals, providing a robust and reliable method
for attendance tracking. The integration of an Arduino micro-
controller facilitates seamless communication between the face
detection algorithm and the attendance log, which is maintained
on an Excel sheet. One of the key features of the system is
its real-time face detection capability, ensuring prompt and
accurate attendance tracking. The solution also demonstrates a
user-friendly interface, allowing for easy deployment in various
settings, such as classrooms, offices, or events. The system’s
adaptability and flexibility make it a valuable tool for automating
attendance processes in diverse environments. The hardware
setup, consisting of a webcam and Arduino, ensures a cost-
effective and accessible solution without compromising on perfor-
mance. The system’s efficiency is further highlighted by its ability
to maintain an organized and easily accessible attendance log in
the form of an Excel sheet. The Face Detection Attendance System
presented in this project showcases a comprehensive and effective
solution for automating attendance tracking. The combination
of Python and Arduino technologies results in a powerful and
versatile system, demonstrating promising results in terms of
accuracy, efficiency, and real-time monitoring.
Index Terms—face detection, attendance system, Python, Ar-
duino, computer vision